% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_bib.R
\name{flex_bib}
\alias{flex_bib}
\title{flex_bib}
\usage{
flex_bib(
  rmarkdown_file,
  bib_input,
  bib_output,
  by_sections = NULL,
  repair = TRUE,
  removeISSN = TRUE,
  removeISBN = TRUE,
  removeDOI = TRUE,
  removeURL = TRUE
)
}
\arguments{
\item{rmarkdown_file}{RMarkdown file in which bibliography shall be inserted}

\item{bib_input}{A vector of two bib files}

\item{bib_output}{Name of the newly created partial .bib files}
}
\value{
Newly created partial .bib files
}
\description{
Creates new output bib files that only contain the cited
   works within predefined reference sections
}
